+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Support Show AI Personality Names mod in place of the built-in war option
+
 ### Changed
 
 - Removed land mine builds from Penchant AI for performance reasons